rename class and add documentation
