+++ Tutorial/src/main/java/programinhas/jokenpo/IAPlayer.java
@@ -1,5 +1,13 @@
 package programinhas.jokenpo;
 
+/**
+ * <h1>IAPlayer</h1>
+ * <p>Classe responsável em criar o jogador <b>IAPlayer</b> através da implementação da interface Player.</p>
+ * <p>Esse jogador possui as jogadas criadas de forma aleatória.</p>
+ * @author Matheus Sproesser
+ * @version 1.0
+ * @since 25/12/2022
+ */
 public class IAPlayer implements Player {
 
     private final String name;
@@ -13,6 +21,10 @@ public class IAPlayer implements Player {
         return this.name;
     }
 
+    /**
+     * Método que irá gerar a escolha do jogador aleatóriamente usando a classe <b>Options</b>.
+     * @return
+     */
     @Override
     public Options getGesture() {
         int index = (int) Math.floor(Math.random() * Options.values().length);
