+++ Tutorial/src/main/java/programinhas/jokenpo/MatchupResult.java
@@ -1,5 +1,13 @@
 package programinhas.jokenpo;
 
+/**
+ * <h1>MatchupResult</h1>
+ * <p>Classe do tipo Enum, responsável por criar as constantes <b>WIN, DRAW e LOSE</b>.</p>
+ * </p>Durante o round, WIN, DRAW E LOSE armazenarão os valores que definirão o resultado do confronto.</p>
+ *  @author Matheus Sproesser
+ *  @version 1.0
+ *  @since 25/12/2022
+ */
 public enum MatchupResult {
     WIN,
     DRAW,
