+++ Tutorial/src/main/java/programinhas/jokenpo/ShellPlayer.java
@@ -3,6 +3,16 @@ package programinhas.jokenpo;
 import java.awt.desktop.SystemSleepEvent;
 import java.util.Scanner;
 
+/**
+ * <h1>ShellPlayer</h1>
+ * <p>Classe responsável em possibilitar entrada de dados pelo usuario.</p>
+ * <p>Nessa modalidade ShellPlayer deverá <b>substituir uma IAPlayer na classe Inicio no método main.</b></p>
+ * @author Matheus Sproesser
+ * @version 1.1
+ * @since 25/12/2022
+ * @contributed Lígia Alves
+ */
+
 public class ShellPlayer implements Player {
 
     private final String name;
@@ -17,6 +27,11 @@ public class ShellPlayer implements Player {
         return name;
     }
 
+    /**
+     * <p>Método responsável em obter a escolha do jogador para confrontar com o seu oponente.</p>
+     * <p>Esse metodo possui um laço de repetição <b>for</b> para verificar se a escolha é menor ou igual ao tamanho do array das constantes da classe Options.</p>
+     * @return a jogada escolhida conforme classe Options.
+     */
     @Override
     public Options getGesture() {
         var options = Options.values();
